RF: replace markers in order they were inserted
The expanded searching meant that we can substitute text with markers,
for text already containing markers.  In that case, the order in which
the text goes back matters, so force order that markers are replaced to
be reverse order to that in which they were inserted.

The test isn't very satisfactory though, and does not fail without the
edit, at least on my machine.