RF: refactor math_dollar into transform
A change in Sphinx exposed a bug in docutils, where the SmartQuotes
transform was throwing away the rawsource of the nodes:

See https://groups.google.com/forum/#!topic/sphinx-dev/v43JedWILR0 for
an analysis.

Avoid this problem by converting math_dollar to a transform, that runs
before the SmartQuotes transform.